requirements_web.txt
flask>=3.0.0
flask-cors>=4.0.0
pypdf>=4.0.0
python-docx>=1.1.0
openai>=1.40.0
python-dotenv>=1.0.0
